Trait necsim_core::cogs::active_lineage_sampler::SingularActiveLineageSampler

pub trait SingularActiveLineageSampler<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry>: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> {
#[must_use]    fn replace_active_lineage(
        &mut self, 
        active_lineage: Option<Lineage>
    ) -> Option<Lineage>;
}

Required methods

#[must_use]fn replace_active_lineage(
&mut self,
active_lineage: Option<Lineage>
) -> Option<Lineage>

Loading content...