Struct necsim_core::cogs::backup::BackedUp

pub struct BackedUp<T: Backup>(/* private fields */);

Implementations

impl<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: LineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, S, X, D, C, T, N>, I: ImmigrationEntry<M>, A: ActiveLineageSampler<M, H, G, S, X, D, C, T, N, E, I>> BackedUp<Simulation<M, H, G, S, X, D, C, T, N, E, I, A>>

pub fn resume(&self) -> Simulation<M, H, G, S, X, D, C, T, N, E, I, A>