Trait necsim_core::cogs::event_sampler::EventSampler

pub trait EventSampler<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: LineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>>: Backup + Debug {
    // Provided method
    fn sample_event_for_lineage_at_event_time_or_emigrate<Q, Aux, FS: FnOnce(SpeciationEvent, Aux) -> Q, FD: FnOnce(DispersalEvent, Aux) -> Q, FE: FnOnce(Aux) -> Q>(
        &mut self,
        lineage: Lineage,
        event_time: PositiveF64,
        simulation: &mut PartialSimulation<M, H, G, S, X, D, C, T, N>,
        rng: &mut G,
        handler: EventHandler<FS, FD, FE>,
        auxiliary: Aux
    ) -> Q { ... }
}

Provided Methods

fn sample_event_for_lineage_at_event_time_or_emigrate<Q, Aux, FS: FnOnce(SpeciationEvent, Aux) -> Q, FD: FnOnce(DispersalEvent, Aux) -> Q, FE: FnOnce(Aux) -> Q>( &mut self, lineage: Lineage, event_time: PositiveF64, simulation: &mut PartialSimulation<M, H, G, S, X, D, C, T, N>, rng: &mut G, handler: EventHandler<FS, FD, FE>, auxiliary: Aux ) -> Q