Trait necsim_core::cogs::LocallyCoherentLineageStore

pub trait LocallyCoherentLineageStore<H: Habitat, R: LineageReference<H>>: LineageStore<H, R> + Index<R, Output = Lineage> {
    fn get_active_global_lineage_reference_at_indexed_location(
        &self, 
        indexed_location: &IndexedLocation, 
        habitat: &H
    ) -> Option<&GlobalLineageReference> { ... }

    fn insert_lineage_to_indexed_location_locally_coherent(
        &mut self, 
        reference: R, 
        indexed_location: IndexedLocation, 
        habitat: &H
    ) { ... }

    fn extract_lineage_from_its_location_locally_coherent(
        &mut self, 
        reference: R, 
        event_time: PositiveF64, 
        habitat: &H
    ) -> (IndexedLocation, NonNegativeF64) { ... }

    fn emigrate(&mut self, local_lineage_reference: R) -> GlobalLineageReference { ... }

    fn immigrate_locally_coherent(
        &mut self, 
        habitat: &H, 
        global_reference: GlobalLineageReference, 
        indexed_location: IndexedLocation, 
        time_of_emigration: PositiveF64
    ) -> R { ... }
}

Provided methods

fn get_active_global_lineage_reference_at_indexed_location(
&self,
indexed_location: &IndexedLocation,
habitat: &H
) -> Option<&GlobalLineageReference>

Contracts

Pre-condition - debug: indexed location is inside habitat

habitat . contains(indexed_location . location())

fn insert_lineage_to_indexed_location_locally_coherent(
&mut self,
reference: R,
indexed_location: IndexedLocation,
habitat: &H
)

Contracts

Pre-condition - debug: indexed location is inside habitat

habitat . contains(indexed_location . location())

Pre-condition - debug: lineage reference is valid

self . get(reference . clone()) . is_some()

Pre-condition - debug: lineage is inactive

! self [reference . clone()] . is_active()

Post-condition - debug: lineage was activated

self [old(reference . clone())] . is_active()

Post-condition - debug: lineage was added to indexed_location

self [old(reference . clone())] . indexed_location() == Some(& old(indexed_location . clone()))

Post-condition - debug: lineage is now indexed at indexed_location

self . get_active_global_lineage_reference_at_indexed_location(& old(indexed_location . clone()), old(habitat)) == Some(self [old(reference . clone())] . global_reference())

fn extract_lineage_from_its_location_locally_coherent(
&mut self,
reference: R,
event_time: PositiveF64,
habitat: &H
) -> (IndexedLocation, NonNegativeF64)

Contracts

Pre-condition - debug: lineage reference is valid

self . get(reference . clone()) . is_some()

Pre-condition - debug: lineage is active

self [reference . clone()] . is_active()

Post-condition - debug: prior location is inside habitat

old(habitat) . contains(ret . 0 . location())

Post-condition - debug: lineage was deactivated

! self [old(reference . clone())] . is_active()

Post-condition - debug: returns the individual’s prior IndexedLocation

ret . 0 == old(self [reference . clone()] . indexed_location() . unwrap() . clone())

Post-condition - debug: returns the individual’s prior last event time

ret . 1 == old(self [reference . clone()] . last_event_time())

Post-condition - debug: lineage is no longer indexed at its prior IndexedLocation

self . get_active_global_lineage_reference_at_indexed_location(& ret . 0, old(habitat)) . is_none()

Post-condition - debug: updates the time of the last event of the lineage reference

self [old(reference . clone())] . last_event_time() == old(event_time)

fn emigrate(&mut self, local_lineage_reference: R) -> GlobalLineageReference

Contracts

Pre-condition - debug: lineage reference is valid

self . get(local_lineage_reference . clone()) . is_some()

Pre-condition - debug: lineage is inactive

! self [local_lineage_reference . clone()] . is_active()

Post-condition - debug: lineage was removed

self . get(old(local_lineage_reference . clone())) . is_none()

Post-condition - debug: returns the individual’s GlobalLineageReference

ret == old(self [local_lineage_reference . clone()] . global_reference() . clone())

fn immigrate_locally_coherent(
&mut self,
habitat: &H,
global_reference: GlobalLineageReference,
indexed_location: IndexedLocation,
time_of_emigration: PositiveF64
) -> R

Contracts

Pre-condition - debug: indexed location is inside habitat

habitat . contains(indexed_location . location())

Post-condition - debug: lineage was activated

self [ret . clone()] . is_active()

Post-condition - debug: lineage was added to indexed_location

self [ret . clone()] . indexed_location() == Some(& old(indexed_location . clone()))

Post-condition - debug: lineage is now indexed at indexed_location

self . get_active_global_lineage_reference_at_indexed_location(& old(indexed_location . clone()), old(habitat)) == Some(self [ret . clone()] . global_reference())

Loading content...