Trait necsim_core::cogs::OptionallyPeekableActiveLineageSampler

pub trait OptionallyPeekableActiveLineageSampler<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry>: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> {
    fn peek_optional_time_of_next_event(
        &mut self, 
        habitat: &H, 
        turnover_rate: &T, 
        rng: &mut G
    ) -> Option<PositiveF64>;
}

Required methods

fn peek_optional_time_of_next_event(
&mut self,
habitat: &H,
turnover_rate: &T,
rng: &mut G
) -> Option<PositiveF64>

Loading content...