Struct necsim_core::simulation::partial::active_lineage_sampler::PartialSimulation

#[repr(C)]
pub struct PartialSimulation<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: LineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, S, X, D, C, T, N>> {
    pub maths: PhantomData<M>,
    pub habitat: H,
    pub lineage_store: S,
    pub dispersal_sampler: D,
    pub coalescence_sampler: C,
    pub turnover_rate: T,
    pub speciation_probability: N,
    pub emigration_exit: X,
    pub event_sampler: E,
    /* private fields */
}

Fields
maths: PhantomData<M>habitat: Hlineage_store: Sdispersal_sampler: Dcoalescence_sampler: Cturnover_rate: Tspeciation_probability: Nemigration_exit: Xevent_sampler: E
Implementations

impl<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: LineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, S, X, D, C, T, N>> PartialSimulation<M, H, G, S, X, D, C, T, N, E>

pub fn with_split_event_sampler<Q, F: FnOnce(&E, &PartialSimulation<M, H, G, S, X, D, C, T, N>) -> Q>(
&self,
func: F
) -> Q

pub fn with_mut_split_event_sampler<Q, F: FnOnce(&mut E, &mut PartialSimulation<M, H, G, S, X, D, C, T, N>) -> Q>(
&mut self,
func: F
) -> Q