Function necsim_core::simulation::process::local::simulate_and_report_local_step_or_finish

pub fn simulate_and_report_local_step_or_finish<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>, P: Reporter>(
    simulation: &mut Simulation<H, G, R, S, X, D, C, T, N, E, I, A>, 
    reporter: &mut P
) -> bool