Struct necsim_core::simulation::Simulation

#[repr(C)]
pub struct Simulation<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>> { /* fields omitted */ }

Implementations

impl<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>> Simulation<H, G, R, S, X, D, C, T, N, E, I, A>

pub fn builder(
) -> SimulationBuilder<((), (), (), (), (), (), (), (), (), (), (), ()), H, G, R, S, X, D, C, T, N, E, I, A>

Create a builder for building Simulation. On the builder, call .habitat(...), .lineage_reference(...), .lineage_store(...), .dispersal_sampler(...), .coalescence_sampler(...), .turnover_rate(...), .speciation_probability(...), .emigration_exit(...), .event_sampler(...), .active_lineage_sampler(...), .rng(...), .immigration_entry(...) to set the values of the fields. Finally, call .build() to create the instance of Simulation.

impl<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>> Simulation<H, G, R, S, X, D, C, T, N, E, I, A>

pub fn with_mut_split_active_lineage_sampler_and_rng<Q, F: FnOnce(&mut A, &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, &mut G) -> Q>(
&mut self,
func: F
) -> Q

pub fn with_mut_split_event_sampler_and_rng<Q, F: FnOnce(&mut E, &PartialSimulation<H, G, R, S, X, D, C, T, N>, &mut G) -> Q>(
&mut self,
func: F
) -> Q

pub fn rng_mut(&mut self) -> &mut G

pub fn active_lineage_sampler(&self) -> &A

pub fn active_lineage_sampler_mut(&mut self) -> &mut A

pub fn lineage_store(&self) -> &S

pub fn lineage_store_mut(&mut self) -> &mut S

pub fn event_sampler(&self) -> &E

pub fn event_sampler_mut(&mut self) -> &mut E

pub fn speciation_probability(&self) -> &N

pub fn turnover_rate(&self) -> &T

pub fn habitat(&self) -> &H

pub fn dispersal_sampler(&self) -> &D

pub fn coalescence_sampler(&self) -> &C

pub fn emigration_exit(&self) -> &X

pub fn emigration_exit_mut(&mut self) -> &mut X

pub fn immigration_entry_mut(&mut self) -> &mut I

impl<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: PeekableActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>> Simulation<H, G, R, S, X, D, C, T, N, E, I, A>

pub fn peek_time_of_next_event(&mut self) -> Option<PositiveF64>

impl<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry, A: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I>> Simulation<H, G, R, S, X, D, C, T, N, E, I, A>

pub fn get_balanced_remaining_work(&self) -> Wrapping<u64>

pub fn simulate_incremental_early_stop<F: FnMut(&mut Self, u64) -> bool, P: Reporter>(
&mut self,
early_stop: F,
reporter: &mut P
) -> (NonNegativeF64, u64)

pub fn simulate<P: Reporter>(self, reporter: &mut P) -> (NonNegativeF64, u64, G)