Trait necsim_core::cogs::lineage_store::GloballyCoherentLineageStore

pub trait GloballyCoherentLineageStore<M: MathsCore, H: Habitat<M>, R: LineageReference<M, H>>: LocallyCoherentLineageStore<M, H, R> {
    type LocationIterator: Iterator<Item = Location>;
    fn iter_active_locations(&self, habitat: &H) -> Self::LocationIterator { ... }

    fn get_local_lineage_references_at_location_unordered(
        &self, 
        location: &Location, 
        habitat: &H
    ) -> &[R] { ... }

    fn insert_lineage_globally_coherent(
        &mut self, 
        lineage: Lineage, 
        habitat: &H
    ) -> R { ... }

    fn extract_lineage_globally_coherent(
        &mut self, 
        reference: R, 
        habitat: &H
    ) -> Lineage { ... }
}

Associated Types

type LocationIterator: Iterator<Item = Location>

Provided methods

fn iter_active_locations(&self, habitat: &H) -> Self::LocationIterator

fn get_local_lineage_references_at_location_unordered(
&self,
location: &Location,
habitat: &H
) -> &[R]

Contracts

Pre-condition - debug: location is inside habitat

habitat.contains(location)

fn insert_lineage_globally_coherent(
&mut self,
lineage: Lineage,
habitat: &H
) -> R

Contracts

Post-condition - debug: lineage is now indexed unordered at indexed_location.location()

self.get_local_lineage_references_at_location_unordered(& old(lineage.indexed_location.location().clone()), old(habitat)).last() == Some(& ret)

Post-condition - debug: unordered active lineage index at given location has grown by 1

`old(self.get_local_lineage_references_at_location_unordered(lineage.indexed_location.location(), old(habitat)).len()

== self.get_local_lineage_references_at_location_unordered(& old(lineage.indexed_location.location().clone()), old(habitat)).len()`

fn extract_lineage_globally_coherent(
&mut self,
reference: R,
habitat: &H
) -> Lineage

Contracts

Post-condition - debug: unordered active lineage index at returned location has shrunk by 1

self.get_local_lineage_references_at_location_unordered(ret.indexed_location.location(), old(habitat),).len() + 1 == old(self.get_local_lineage_references_at_location_unordered(self [reference.clone()].indexed_location.location(), old(habitat),).len())