Trait necsim_core::cogs::ActiveLineageSampler

pub trait ActiveLineageSampler<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry>: Backup + Debug {
    fn number_active_lineages(&self) -> usize { ... }

    fn get_last_event_time(&self) -> NonNegativeF64 { ... }

    fn pop_active_lineage_indexed_location_prior_event_time(
        &mut self, 
        simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, 
        rng: &mut G
    ) -> Option<(R, IndexedLocation, NonNegativeF64, PositiveF64)> { ... }

    fn push_active_lineage_to_indexed_location(
        &mut self, 
        lineage_reference: R, 
        indexed_location: IndexedLocation, 
        time: PositiveF64, 
        simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, 
        rng: &mut G
    ) { ... }

    fn insert_new_lineage_to_indexed_location(
        &mut self, 
        global_reference: GlobalLineageReference, 
        indexed_location: IndexedLocation, 
        time: PositiveF64, 
        simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, 
        rng: &mut G
    ) { ... }

    fn with_next_active_lineage_indexed_location_prior_event_time<F: FnOnce(&mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, &mut G, R, IndexedLocation, NonNegativeF64, PositiveF64) -> Option<IndexedLocation>>(
        &mut self, 
        simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, 
        rng: &mut G, 
        inner: F
    ) -> bool { ... }
}

Provided methods

fn number_active_lineages(&self) -> usize

fn get_last_event_time(&self) -> NonNegativeF64

fn pop_active_lineage_indexed_location_prior_event_time(
&mut self,
simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>,
rng: &mut G
) -> Option<(R, IndexedLocation, NonNegativeF64, PositiveF64)>

Contracts

Post-condition - debug: removes an active lineage if some left

match ret { Some(_) => { self.number_active_lineages() == old(self.number_active_lineages()) - 1 }, None => old(self.number_active_lineages()) == 0, }

Post-condition - debug: event occurs later than last event time

ret.is_some() -> ret.as_ref().unwrap().3 > old(self.get_last_event_time())

Post-condition - debug: updates the time of the last event of the returned lineage to event time

if let Some((ref reference, ref _location, _prior_time, event_time)) = ret { simulation.lineage_store.get(reference.clone()).map_or(true, | lineage | { lineage.last_event_time() == event_time }) } else { true }

Post-condition - debug: updates the time of the last event

if let Some((ref _reference, ref _location, _prior_time, event_time)) = ret { self.get_last_event_time() == event_time } else { true }

fn push_active_lineage_to_indexed_location(
&mut self,
lineage_reference: R,
indexed_location: IndexedLocation,
time: PositiveF64,
simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>,
rng: &mut G
)

Contracts

Post-condition - debug: adds an active lineage

self.number_active_lineages() == old(self.number_active_lineages()) + 1

fn insert_new_lineage_to_indexed_location(
&mut self,
global_reference: GlobalLineageReference,
indexed_location: IndexedLocation,
time: PositiveF64,
simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>,
rng: &mut G
)

Contracts

Post-condition - debug: adds an active lineage

self.number_active_lineages() == old(self.number_active_lineages()) + 1

fn with_next_active_lineage_indexed_location_prior_event_time<F: FnOnce(&mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>, &mut G, R, IndexedLocation, NonNegativeF64, PositiveF64) -> Option<IndexedLocation>>(
&mut self,
simulation: &mut PartialSimulation<H, G, R, S, X, D, C, T, N, E>,
rng: &mut G,
inner: F
) -> bool