Trait necsim_core::cogs::GloballyCoherentLineageStore

pub trait GloballyCoherentLineageStore<H: Habitat, R: LineageReference<H>>: LocallyCoherentLineageStore<H, R> {
    type LocationIterator: Iterator<Item = Location>;
    fn iter_active_locations(&self, habitat: &H) -> Self::LocationIterator { ... }

    fn get_active_local_lineage_references_at_location_unordered(
        &self, 
        location: &Location, 
        habitat: &H
    ) -> &[R] { ... }

    fn insert_lineage_to_indexed_location_globally_coherent(
        &mut self, 
        reference: R, 
        indexed_location: IndexedLocation, 
        habitat: &H
    ) { ... }

    fn extract_lineage_from_its_location_globally_coherent(
        &mut self, 
        reference: R, 
        event_time: PositiveF64, 
        habitat: &H
    ) -> (IndexedLocation, NonNegativeF64) { ... }

    fn immigrate_globally_coherent(
        &mut self, 
        habitat: &H, 
        global_reference: GlobalLineageReference, 
        indexed_location: IndexedLocation, 
        time_of_emigration: PositiveF64
    ) -> R { ... }
}

Associated Types

type LocationIterator: Iterator<Item = Location>

Provided methods

fn iter_active_locations(&self, habitat: &H) -> Self::LocationIterator

fn get_active_local_lineage_references_at_location_unordered(
&self,
location: &Location,
habitat: &H
) -> &[R]

Contracts

Pre-condition - debug: location is inside habitat

habitat.contains(location)

fn insert_lineage_to_indexed_location_globally_coherent(
&mut self,
reference: R,
indexed_location: IndexedLocation,
habitat: &H
)

Contracts

Post-condition - debug: lineage is now indexed unordered at indexed_location.location()

self.get_active_local_lineage_references_at_location_unordered(& old(indexed_location.location().clone()), old(habitat)).last() == Some(& old(reference.clone()))

Post-condition - debug: unordered active lineage index at given location has grown by 1

`old(self.get_active_local_lineage_references_at_location_unordered(indexed_location.location(), old(habitat)).len()

== self.get_active_local_lineage_references_at_location_unordered(& old(indexed_location.location().clone()), old(habitat)).len()`

fn extract_lineage_from_its_location_globally_coherent(
&mut self,
reference: R,
event_time: PositiveF64,
habitat: &H
) -> (IndexedLocation, NonNegativeF64)

Contracts

Post-condition - debug: unordered active lineage index at returned location has shrunk by 1

`self.get_active_local_lineage_references_at_location_unordered(ret.0.location(), old(habitat),).len()

1 == old(self.get_active_local_lineage_references_at_location_unordered(self [reference.clone()].indexed_location().unwrap().location(), old(habitat),).len())`

fn immigrate_globally_coherent(
&mut self,
habitat: &H,
global_reference: GlobalLineageReference,
indexed_location: IndexedLocation,
time_of_emigration: PositiveF64
) -> R

Contracts

Post-condition - debug: lineage is now indexed unordered at indexed_location.location()

self.get_active_local_lineage_references_at_location_unordered(& old(indexed_location.location().clone()), old(habitat)).last() == Some(& ret)

Post-condition - debug: unordered active lineage index at given location has grown by 1

`old(self.get_active_local_lineage_references_at_location_unordered(indexed_location.location(), old(habitat)).len()

== self.get_active_local_lineage_references_at_location_unordered(& old(indexed_location.location().clone()), old(habitat)).len()`