Trait necsim_core::cogs::PeekableActiveLineageSampler

pub trait PeekableActiveLineageSampler<H: Habitat, G: RngCore, R: LineageReference<H>, S: LineageStore<H, R>, X: EmigrationExit<H, G, R, S>, D: DispersalSampler<H, G>, C: CoalescenceSampler<H, R, S>, T: TurnoverRate<H>, N: SpeciationProbability<H>, E: EventSampler<H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry>: ActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> {
    fn peek_time_of_next_event(
        &mut self, 
        habitat: &H, 
        turnover_rate: &T, 
        rng: &mut G
    ) -> Result<PositiveF64, EmptyActiveLineageSamplerError> { ... }
}

Provided methods

fn peek_time_of_next_event(
&mut self,
habitat: &H,
turnover_rate: &T,
rng: &mut G
) -> Result<PositiveF64, EmptyActiveLineageSamplerError>

Contracts

Post-condition - debug: only returns Err when no more lineages remain

ret . is_err() == (self . number_active_lineages() == 0)

Loading content...