Struct necsim_core::simulation::Simulation

#[repr(C)]
pub struct Simulation<M: MathsCore, H: Habitat<M>, G: RngCore<M>, R: LineageReference<M, H>, S: LineageStore<M, H, R>, X: EmigrationExit<M, H, G, R, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, R, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry<M>, A: ActiveLineageSampler<M, H, G, R, S, X, D, C, T, N, E, I>> { /* fields omitted */ }

Implementations

impl<M: MathsCore, H: Habitat<M>, G: RngCore<M>, R: LineageReference<M, H>, S: LineageStore<M, H, R>, X: EmigrationExit<M, H, G, R, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, R, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry<M>, A: ActiveLineageSampler<M, H, G, R, S, X, D, C, T, N, E, I>> Simulation<M, H, G, R, S, X, D, C, T, N, E, I, A>

pub fn with_mut_split_active_lineage_sampler_and_rng<Q, F: FnOnce(&mut A, &mut PartialSimulation<M, H, G, R, S, X, D, C, T, N, E>, &mut G) -> Q>(
&mut self,
func: F
) -> Q

pub fn with_mut_split_event_sampler_and_rng<Q, F: FnOnce(&mut E, &PartialSimulation<M, H, G, R, S, X, D, C, T, N>, &mut G) -> Q>(
&mut self,
func: F
) -> Q

pub fn rng_mut(&mut self) -> &mut G

pub fn active_lineage_sampler(&self) -> &A

pub fn active_lineage_sampler_mut(&mut self) -> &mut A

pub fn lineage_store(&self) -> &S

pub fn lineage_store_mut(&mut self) -> &mut S

pub fn event_sampler(&self) -> &E

pub fn event_sampler_mut(&mut self) -> &mut E

pub fn speciation_probability(&self) -> &N

pub fn turnover_rate(&self) -> &T

pub fn habitat(&self) -> &H

pub fn dispersal_sampler(&self) -> &D

pub fn coalescence_sampler(&self) -> &C

pub fn emigration_exit(&self) -> &X

pub fn emigration_exit_mut(&mut self) -> &mut X

pub fn immigration_entry_mut(&mut self) -> &mut I

impl<M: MathsCore, H: Habitat<M>, G: RngCore<M>, R: LineageReference<M, H>, S: LineageStore<M, H, R>, X: EmigrationExit<M, H, G, R, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, R, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry<M>, A: ActiveLineageSampler<M, H, G, R, S, X, D, C, T, N, E, I>> Simulation<M, H, G, R, S, X, D, C, T, N, E, I, A>

pub fn is_done(&self) -> bool

pub fn get_balanced_remaining_work(&self) -> Wrapping<u64>

pub fn simulate_incremental_early_stop<F: FnMut(&Self, u64, PositiveF64) -> bool, P: Reporter>(
&mut self,
early_stop: F,
reporter: &mut P
) -> (NonNegativeF64, u64)

pub fn simulate<P: Reporter>(self, reporter: &mut P) -> (NonNegativeF64, u64)